requirements.txt
aiogram<3.0
environs~=8.0.0
asyncpg-0.27.0